+++ km3pipe/utils/tohdf5.py
@@ -16,6 +16,8 @@ Options:
     --aa-lib-<lib.so>          (Aanet): path to aanet binary (for old versions which
                                must be loaded via `ROOT.gSystem.Load()` instead
                                of `import aa`)
+    --aa-old-mc-id             (aanet): read mc id as `evt.mc_id`, instead
+                               of the newer `mc_id = evt.frame_index - 1`
     -h --help                  Show this screen.
     -j --jppy                  (Jpp): Use jppy (not aanet) for Jpp readout
     -l --with-timeslice-hits   Include timeslice-hits [default: False]
@@ -86,6 +88,7 @@ def main():
     use_jppy_pump = args['--jppy']
     aa_format = args['--aa-format']
     aa_lib = args['--aa-lib']
+    aa_old_mc_id = args['--aa-old-mc-id']
     with_summaryslices = args['--with-summaryslices']
     with_timeslice_hits = args['--with-timeslice-hits']
     correct_zed = args['--correct-zed']
@@ -99,4 +102,5 @@ def main():
            with_timeslice_hits=with_timeslice_hits,
            n_rows_expected=n_rows_expected,
            apply_zed_correction=correct_zed,
+           old_mc_id=aa_old_mc_id,
            )